# docs/swagger.yaml
basePath: /api
definitions:
  auth.LoginDto:
    properties:
      password:
        maxLength: 200
        minLength: 1
        type: string
      username:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  category.CategoryDto:
    properties:
      id:
        maximum: 10
        minimum: 1
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - name
    type: object
  product.ProductDto:
    properties:
      id:
        maximum: 10
        minimum: 1
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
      price:
        maximum: 11
        minimum: 1
        type: number
    required:
    - name
    - price
    type: object
  user.UserDto:
    properties:
      id:
        maximum: 10
        minimum: 1
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
      password:
        maxLength: 200
        minLength: 1
        type: string
      username:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - name
    - password
    - username
    type: object
host: localhost:8899
info:
  contact: {}
  description: How to implement swagger-ui in Go and gin http client project
  title: Nexsoft Demo Swagger in GO (gin)
  version: 1.1.0
paths:
  /v1/auth/login:
    post:
      description: Get JWT token for access all APIs
      parameters:
      - description: LoginDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: login
      tags:
      - auth
  /v1/category/delete/{id}:
    delete:
      description: delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (CategoryDto.go)
          schema:
            $ref: '#/definitions/category.CategoryDto'
      summary: delete
      tags:
      - category
  /v1/category/list:
    get:
      description: list
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (CategoryDto.go)
          schema:
            $ref: '#/definitions/category.CategoryDto'
      summary: list
      tags:
      - category
  /v1/category/save:
    post:
      description: save
      parameters:
      - description: CategoryDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/category.CategoryDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (CategoryDto.go)
          schema:
            $ref: '#/definitions/category.CategoryDto'
      summary: save
      tags:
      - category
  /v1/category/update/{id}:
    put:
      description: update
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: CategoryDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/category.CategoryDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (CategoryDto.go)
          schema:
            $ref: '#/definitions/category.CategoryDto'
      summary: update
      tags:
      - category
  /v1/category/view/{id}:
    get:
      description: view
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (CategoryDto.go)
          schema:
            $ref: '#/definitions/category.CategoryDto'
      summary: view
      tags:
      - category
  /v1/product/delete/{id}:
    delete:
      description: delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (ProductDto.go)
          schema:
            $ref: '#/definitions/product.ProductDto'
      summary: delete
      tags:
      - product
  /v1/product/list:
    get:
      description: list
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (ProductDto.go)
          schema:
            $ref: '#/definitions/product.ProductDto'
      summary: list
      tags:
      - product
  /v1/product/save:
    post:
      description: save
      parameters:
      - description: ProductDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/product.ProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (ProductDto.go)
          schema:
            $ref: '#/definitions/product.ProductDto'
      summary: save
      tags:
      - product
  /v1/product/update/{id}:
    put:
      description: update
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: ProductDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/product.ProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (ProductDto.go)
          schema:
            $ref: '#/definitions/product.ProductDto'
      summary: update
      tags:
      - product
  /v1/product/view/{id}:
    get:
      description: view
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (ProductDto.go)
          schema:
            $ref: '#/definitions/product.ProductDto'
      summary: view
      tags:
      - product
  /v1/user/delete/{id}:
    delete:
      description: delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: delete
      tags:
      - user
  /v1/user/list:
    get:
      description: list
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: list
      tags:
      - user
  /v1/user/save:
    post:
      description: save
      parameters:
      - description: UserDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/user.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: save
      tags:
      - user
  /v1/user/update/{id}:
    put:
      description: update
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: UserDto.go
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/user.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: update
      tags:
      - user
  /v1/user/view/{id}:
    get:
      description: view
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response Success (UserDto.go)
          schema:
            $ref: '#/definitions/user.UserDto'
      summary: view
      tags:
      - user
swagger: "2.0"